+++ src/hir/cache_codec.rs
@@ -90,7 +90,7 @@ codec_struct!(ResolvedNativeRustImportCall {
     args,
     result
 });
-codec_enum!(ResolvedHostCommandOperation {0=>ArgsLen,1=>ArgUtf8,2=>StdinRead,3=>StderrWrite,4=>StdoutAppend,5=>StderrAppend,6=>NetConnect,7=>NetSend,8=>NetRecv,9=>NetStreamStdout,10=>NetWait,11=>NetClose});
+codec_enum!(ResolvedHostCommandOperation {0=>ArgsLen,1=>ArgUtf8,2=>StdinRead,3=>StderrWrite,4=>StdoutAppend,5=>StderrAppend,6=>NetConnect,7=>NetSend,8=>NetRecv,9=>NetStreamStdout,10=>NetWait,11=>NetClose,12=>NetTlsConnect,13=>NetListen,14=>NetAccept,15=>NetCloseListener});
 codec_struct!(ResolvedHostCommandCall {
     expression,
     operation,